requirements.txt
Django==1.10.5
PyYAML==3.12
argparse==1.2.1
django-filter==1.0.1
djangorestframework==3.5.3
psycopg2==2.6.2
pyPdf==1.13
requests==2.13.0
slugify==0.0.1
wsgiref==0.1.2
